improve mode inference implementation, Consolidate inference logic, simplify validation, and reorganize tests based on review
